added eightball command
