+++ src/bot_abstractions/bot_messages.py
@@ -1,0 +1,56 @@
+class BotMessages:
+    NOT_IN_VC = 'You are not in a voice channel'
+    JOINING_VC = "Joining Voice Channel..."
+    ADDED_TO_QUEUE = "has been added to queue"
+    QUEUE_CLEARED = 'Queue has been cleared'
+    EMPTY_QUEUE = 'There are no songs in the queue'
+    NOW_PLAYING = 'Now Playing: '
+    HELLO_STRINGS = [
+        'Shut up fookin loser lmao',
+        'I did ur mom',
+        'Hey... No one likes you',
+        'You have no friends',
+        'Word has it that you have a small peepee',
+        'Dont you have anything better to do than say hi to me?',
+        'Fill in the blank: _____ is a bad Astra Player. What are you doing with ur late lurks?',
+        'Fill in the blank: _____ is a bad Killjoy Player that likes to get railed from the side',
+        'Fill in the blank: _____ is a bad Jett Player',
+        'Fill in the blank: _____ is a bad Sage Player, you wish you had mommy milkers like her',
+        'Fill in the blank: _____ is a bad Harbor Player',
+        'Hi',
+        'Shut up before I hit the griddy on you',
+        'L + Ratiod',
+        'Shut up you suck',
+        'BROOO SHUDDDAPPPP',
+        'BARK BARK BARK BARK',
+        'Meow',
+        'If I had a dollar for every brain that you didnt have, id have 1 dollar',
+        'I hope you demote in valorant'
+    ]
+
+class EightBallResponses:
+    YES_STRINGS = [
+        'Hell yea',
+        'Ong breh',
+        'Thats a yes for me',
+        'Oh forsure',
+        'Yes',
+        'Ye',
+        'Yeah',
+        'Is that even a question? Yes',
+        'Most def',
+        'Without a doubt',
+        'It is certain'
+    ]
+    NO_STRINGS = [
+        'Dont count on it',
+        'Hell na',
+        'Nope',
+        'Nah breh',
+        'Doubt it'
+    ]
+    MAYBE_STRINGS = [
+        'Ask againe later',
+        'Maybe',
+        'Try Again'
+    ]
